+++ geotrek/api/v2/filters.py
@@ -1,5 +1,6 @@
 from __future__ import unicode_literals
 
+from datetime import date
 import operator
 from functools import reduce
 
@@ -68,13 +69,14 @@ class GeotrekPublishedFilter(BaseFilterBackend):
 
     def filter_queryset(self, request, queryset, view):
         qs = queryset
-        published = request.GET.get('published', None)
+        published = request.GET.get('published', 'true')
 
         if published == 'true':
             published = True
-
         elif published == 'false':
             published = False
+        else:
+            published = None
 
         if published is not None:
             language = request.GET.get('language', 'all')
@@ -104,7 +106,41 @@ class GeotrekPublishedFilter(BaseFilterBackend):
 
     def get_schema_fields(self, view):
         field_published = Field(name='published', required=False,
-                                description=_('Publication state. If language specified, only language published are filterted. true / false'),
+                                description=_('Publication state. If language specified, only language published are filterted. true/false/all. true by default.'),
                                 type='boolean',
                                 example='true')
         return field_published,
+
+
+class GeotrekSensitiveAreaFilter(BaseFilterBackend):
+    def filter_queryset(self, request, queryset, view):
+        qs = queryset
+        practices = request.GET.get('practices', '')
+        if practices:
+            qs = qs.filter(species__practices__name__in=practices.split(','))
+        structure = request.GET.get('structure', '')
+        if structure:
+            qs = qs.filter(structure__name=structure)
+        period = request.GET.get('period', '')
+        if not period:
+            qs = qs.filter(**{'species__period{:02}'.format(date.today().month): True})
+        elif period == 'any':
+            qs = qs.filter(reduce(operator.or_, (Q(**{'species__period{:02}'.format(m): True}) for m in range(1, 13))))
+        elif period == 'ignore':
+            pass
+        else:
+            months = [int(m) for m in period.split(',')]
+            qs = qs.filter(reduce(operator.or_, (Q(**{'species__period{:02}'.format(m): True}) for m in months)))
+        return qs
+
+    def get_schema_fields(self, view):
+        field_period = Field(name='period', required=False,
+                             description=_('Period of occupancy. Month numbers (1-12) separated by comas. any = occupied at any time in the year. ignore = occupied or not.'),
+                             example=u'7,8 for july and august')
+        field_practices = Field(name='practices', required=False,
+                                description=_('Practices names separated by comas.'),
+                                example=u'Terrestre,Vertical')
+        field_structure = Field(name='structure', required=False,
+                                description=_('Structure name.'),
+                                example=u'PNM')
+        return field_period, field_practices, field_structure