+++ geotrek/api/v2/functions.py
@@ -1,14 +1,29 @@
 from __future__ import unicode_literals
 
-from django.db.models import Func, F
-from django.db.models.fields import FloatField
+from django.db.models import Func
+from django.db.models.fields import FloatField, CharField
+from django.contrib.gis.db.models import GeometryField
 
 
-def Transform(field_name, srid):
+def Transform(geom, srid):
     """
     ST_TRANSFORM postgis function
     """
-    return Func(F(field_name), srid, function='ST_TRANSFORM')
+    return Func(geom, srid, function='ST_TRANSFORM')
+
+
+def Buffer(geom, radius, num_seg):
+    """
+    ST_Buffer postgis function
+    """
+    return Func(geom, radius, num_seg, function='ST_Buffer', output_field=GeometryField())
+
+
+def GeometryType(geom):
+    """
+    GeometryType postgis function
+    """
+    return Func(geom, function='GeometryType', output_field=CharField())
 
 
 class Length(Func):