+++ geotrek/sensitivity/migrations/0007_species_radius.py
@@ -1,0 +1,19 @@
+# -*- coding: utf-8 -*-
+from __future__ import unicode_literals
+
+from django.db import migrations, models
+
+
+class Migration(migrations.Migration):
+
+    dependencies = [
+        ('sensitivity', '0006_auto_20171016_1422'),
+    ]
+
+    operations = [
+        migrations.AddField(
+            model_name='species',
+            name='radius',
+            field=models.IntegerField(null=True, blank=True),
+        ),
+    ]