+++ geotrek/urls.py
@@ -1,5 +1,5 @@
 from django.conf import settings
-from django.conf.urls import patterns, include, url, static
+from django.conf.urls import include, url, static
 from django.contrib.staticfiles.urls import staticfiles_urlpatterns
 from django.contrib import admin
 
@@ -10,9 +10,7 @@ handler404 = 'mapentity.views.handler404'
 handler500 = 'mapentity.views.handler500'
 
 
-urlpatterns = patterns(
-    '',
-    url(r'^api/v2/', include('geotrek.api.v2.urls', namespace='apiv2', app_name='apiv2')),
+urlpatterns = [
     url(r'^$', 'geotrek.core.views.home', name='home'),
     url(r'^login/$', 'django.contrib.auth.views.login', name='login'),
     url(r'^logout/$', 'django.contrib.auth.views.logout', {'next_page': settings.ROOT_URL + '/'}, name='logout',),
@@ -20,16 +18,6 @@ urlpatterns = patterns(
     url(r'', include('geotrek.common.urls', namespace='common', app_name='common')),
     url(r'', include('geotrek.core.urls', namespace='core', app_name='core')),
     url(r'', include('geotrek.altimetry.urls', namespace='altimetry', app_name='altimetry')),
-    url(r'', include('geotrek.land.urls', namespace='land', app_name='land')),
-    url(r'', include('geotrek.zoning.urls', namespace='zoning', app_name='zoning')),
-    url(r'', include('geotrek.infrastructure.urls', namespace='infrastructure', app_name='infrastructure')),
-    url(r'', include('geotrek.maintenance.urls', namespace='maintenance', app_name='maintenance')),
-    url(r'', include('geotrek.trekking.urls', namespace='trekking', app_name='trekking')),
-    url(r'', include('geotrek.tourism.urls', namespace='tourism', app_name='tourism')),
-    url(r'', include('geotrek.flatpages.urls', namespace='flatpages', app_name='flatpages')),
-    url(r'', include('geotrek.feedback.urls', namespace='feedback', app_name='feedback')),
-
-
 
     url(r'', include('mapentity.urls', namespace='mapentity', app_name='mapentity')),
     url(r'^paperclip/add-for/(?P<app_label>[\w\-]+)/(?P<model_name>[\w\-]+)/(?P<pk>\d+)/$',
@@ -39,7 +27,28 @@ urlpatterns = patterns(
     url(r'^paperclip/', include('paperclip.urls')),
     url(r'^admin/doc/', include('django.contrib.admindocs.urls')),
     url(r'^admin/', include(admin.site.urls)),
-)
+]
+
+if 'geotrek.land' in settings.INSTALLED_APPS:
+    urlpatterns.append(url(r'', include('geotrek.land.urls', namespace='land', app_name='land')))
+if 'geotrek.zoning' in settings.INSTALLED_APPS:
+    urlpatterns.append(url(r'', include('geotrek.zoning.urls', namespace='zoning', app_name='zoning')))
+if 'geotrek.infrastructure' in settings.INSTALLED_APPS:
+    urlpatterns.append(url(r'', include('geotrek.infrastructure.urls', namespace='infrastructure', app_name='infrastructure')))
+if 'geotrek.maintenance' in settings.INSTALLED_APPS:
+    urlpatterns.append(url(r'', include('geotrek.maintenance.urls', namespace='maintenance', app_name='maintenance')))
+if 'geotrek.trekking' in settings.INSTALLED_APPS:
+    urlpatterns.append(url(r'', include('geotrek.trekking.urls', namespace='trekking', app_name='trekking')))
+if 'geotrek.tourism' in settings.INSTALLED_APPS:
+    urlpatterns.append(url(r'', include('geotrek.tourism.urls', namespace='tourism', app_name='tourism')))
+if 'geotrek.flatpages' in settings.INSTALLED_APPS:
+    urlpatterns.append(url(r'', include('geotrek.flatpages.urls', namespace='flatpages', app_name='flatpages')))
+if 'geotrek.feedback' in settings.INSTALLED_APPS:
+    urlpatterns.append(url(r'', include('geotrek.feedback.urls', namespace='feedback', app_name='feedback')))
+if 'geotrek.sensitivity' in settings.INSTALLED_APPS:
+    urlpatterns.append(url(r'', include('geotrek.sensitivity.urls', namespace='sensitivity', app_name='sensitivity')))
+if 'geotrek.api' in settings.INSTALLED_APPS:
+    urlpatterns.append(url(r'^api/v2/', include('geotrek.api.v2.urls', namespace='apiv2', app_name='apiv2')))
 
 urlpatterns += staticfiles_urlpatterns()
 urlpatterns += static.static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)